+++ src/file.h
@@ -3,8 +3,8 @@
 
 namespace File
 {
-    bool Load(const char* fileName, char*& buffer, int& size);
-    bool Save(const char* fileName, const char* buffer, int size);
+    bool Load(const char* fileName, char*& buffer, size_t& size);
+    bool Save(const char* fileName, const char* buffer, size_t size);
 }
 
 #endif // FILE_H_
